requirements.txt
unstructured[all-docs]>=0.10.30
python-docx>=0.8.11
libmagic>=0.4.27 